places recommendation integrated with the recommendation system(python) tested with bff

+++ backend/src/main/java/com/wandr/backend/dto/traveller/UpdateActivitiesDTO.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 @Data
 public class UpdateActivitiesDTO {
-    private List<Integer> activities;
+    private List<Long> activities;
 }
